Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -18,6 +18,6 @@ float f;
 	printf("size of an int: %ld byte(s)\n", sizeof(int));
 	printf("size of a long int: %ld byte(s)\n", sizeof(float));
 	printf("size of a long long int: %ld byte(s)\n", sizeof(double));
-	printf("size of a float %ld byte(s)\n", sizeof(long long int));
+	printf("size of a float %ld byte(s)\n", sizeof(float));
 	return (0);
 }
